После рефакторинга
