+++ src/widgets/appBar/AppBar.tsx
@@ -4,14 +4,13 @@ import UserNotLogin from 'widgets/appBar/UserNotLogin';
 import UserLogin from 'widgets/appBar/UserLogin';
 import { GlobalContext } from 'shared/context/context';
 
-//
-function AppBar(): React.ReactElement {
-    const context = React.useContext<GlobalContext>(Context);
 
+function AppBar(): React.ReactElement {
+    const { isLogin } = React.useContext<GlobalContext>(Context);
 
     return (
         <nav className="flex  max-w-full  justify-center items-start md:py-1 min-w-[50px] lg:min-w-[100px] bg-slate-900 rounded-md shadow-lg  ">
-            {context.isLogin ? <UserLogin /> : <UserNotLogin />}
+            {isLogin ? <UserLogin /> : <UserNotLogin />}
         </nav>
     );
 }
