--- src/widgets/categories/interfaces/interfaces.ts
@@ -1,4 +1,0 @@
-export interface ICategorieItem {
-  categorie: string;
-  id: string;
-}
